fix: resolve login authentication failures and production hardening
- Fix DataLoader: use correct passwords (admin123/staff123) and create users on fresh DB
- Fix frontend authService: read response.data.user instead of response.data.userInfo
- Fix CorsConfig: use CorsConfigurationSource bean instead of WebMvcConfigurer for Spring Security compatibility
- Add BadCredentialsException handler returning 401 instead of 500
- Enable .cors(withDefaults()) in SecurityConfig filter chain
- Remove duplicate ApiResponse, PasswordDebugController, TestController
- Replace wildcard imports with explicit imports in controllers/services
- Add @Transactional annotations to service methods
- Change ChitCycle financial fields from Double to BigDecimal
- Set FetchType.LAZY on ManyToOne relationships
- Add .env for frontend API URL configuration

+++ backend/src/main/java/com/chitfund/backend/dto/ChitCycleDTO.java
@@ -1,6 +1,7 @@
 package com.chitfund.backend.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -10,10 +11,10 @@ public class ChitCycleDTO {
     private LocalDate auctionDate;
     private Long winnerId; // ID of the winner customer
     private String winnerName; // Name of the winner
-    private Double chitPayoutAmount;
-    private Double auctionAmount;
-    private Double commissionAmount;
-    private Double dividendAmount;
-    private Double finalMonthlyAmount;
+    private BigDecimal chitPayoutAmount;
+    private BigDecimal auctionAmount;
+    private BigDecimal commissionAmount;
+    private BigDecimal dividendAmount;
+    private BigDecimal finalMonthlyAmount;
     private String status;
 }
